refactor(dto): 하위 패키지 생성
