--- src/main/java/wooteco/subway/dto/SectionRequest.java
@@ -1,29 +1,0 @@
-package wooteco.subway.dto;
-
-public class SectionRequest {
-
-    private Long upStationId;
-    private Long downStationId;
-    private int distance;
-
-    private SectionRequest() {
-    }
-
-    public SectionRequest(final Long upStationId, final Long downStationId, final int distance) {
-        this.upStationId = upStationId;
-        this.downStationId = downStationId;
-        this.distance = distance;
-    }
-
-    public Long getUpStationId() {
-        return upStationId;
-    }
-
-    public Long getDownStationId() {
-        return downStationId;
-    }
-
-    public int getDistance() {
-        return distance;
-    }
-}
