+++ src/main/java/wooteco/subway/dto/line/LineRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.dto.line;
 
 public class LineRequest {
