+++ src/main/java/wooteco/subway/dto/station/StationRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.dto.station;
 
 public class StationRequest {
