+++ src/test/java/wooteco/subway/acceptance/AcceptanceTest.java
@@ -10,8 +10,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.MediaType;
 import org.springframework.jdbc.core.JdbcTemplate;
-import wooteco.subway.dto.LineRequest;
-import wooteco.subway.dto.StationRequest;
+import wooteco.subway.dto.line.LineRequest;
+import wooteco.subway.dto.station.StationRequest;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class AcceptanceTest {
